+++ zed_components/src/zed_camera/include/sl_types.hpp
@@ -47,7 +47,7 @@
 #include <std_msgs/msg/bool.hpp>
 #include <std_srvs/srv/set_bool.hpp>
 #include <std_srvs/srv/trigger.hpp>
-#include <cob_srvs/srv/set_int.hpp>
+#include <zed_msgs/srv/set_svo_frame.hpp>
 #include <stereo_msgs/msg/disparity_image.hpp>
 #include <string>
 #include <vector>
@@ -58,9 +58,13 @@
 #include <zed_msgs/msg/objects_stamped.hpp>
 #include <zed_msgs/msg/plane_stamped.hpp>
 #include <zed_msgs/msg/pos_track_status.hpp>
+#include <zed_msgs/msg/svo_status.hpp>
+#include <zed_msgs/msg/health_status_stamped.hpp>
+#include <zed_msgs/msg/heartbeat.hpp>
 #include <zed_msgs/srv/set_pose.hpp>
 #include <zed_msgs/srv/set_roi.hpp>
 #include <zed_msgs/srv/start_svo_rec.hpp>
+
 
 #ifndef FOUND_FOXY
   #include <point_cloud_transport/point_cloud_transport.hpp>
@@ -92,6 +96,13 @@ const bool IS_JETSON = false;
 const float NOT_VALID_TEMP = -273.15f;
 
 // ----> Typedefs to simplify declarations
+typedef std::shared_ptr<sensor_msgs::msg::CameraInfo> camInfoMsgPtr;
+typedef std::shared_ptr<rclcpp::Publisher<zed_msgs::msg::SvoStatus>>
+  svoStatusPub;
+typedef std::shared_ptr<rclcpp::Publisher<zed_msgs::msg::HealthStatusStamped>>
+  healthStatusPub;
+typedef std::shared_ptr<rclcpp::Publisher<zed_msgs::msg::Heartbeat>>
+  heartbeatStatusPub;
 
 typedef std::shared_ptr<rclcpp::Publisher<sensor_msgs::msg::Image>> imagePub;
 typedef std::shared_ptr<rclcpp::Publisher<stereo_msgs::msg::DisparityImage>>
@@ -143,35 +154,6 @@ typedef std::shared_ptr<rclcpp::Subscription<geometry_msgs::msg::PointStamped>> 
 typedef std::shared_ptr<rclcpp::Subscription<sensor_msgs::msg::NavSatFix>> gnssFixSub;
 typedef std::shared_ptr<rclcpp::Subscription<rosgraph_msgs::msg::Clock>> clockSub;
 
-//typedef std::unique_ptr<point_cloud_transport::PointCloudTransport> ptTranspPtr;
-
-typedef std::unique_ptr<sensor_msgs::msg::Image> imageMsgPtr;
-typedef std::shared_ptr<sensor_msgs::msg::CameraInfo> camInfoMsgPtr;
-typedef std::unique_ptr<sensor_msgs::msg::PointCloud2> pointcloudMsgPtr;
-typedef std::unique_ptr<sensor_msgs::msg::Imu> imuMsgPtr;
-typedef std::unique_ptr<sensor_msgs::msg::FluidPressure> pressMsgPtr;
-typedef std::unique_ptr<sensor_msgs::msg::Temperature> tempMsgPtr;
-typedef std::unique_ptr<sensor_msgs::msg::MagneticField> magMsgPtr;
-typedef std::unique_ptr<stereo_msgs::msg::DisparityImage> dispMsgPtr;
-
-typedef std::unique_ptr<geometry_msgs::msg::PoseStamped> poseMsgPtr;
-
-typedef std::unique_ptr<zed_msgs::msg::PosTrackStatus> poseStatusMsgPtr;
-typedef std::unique_ptr<zed_msgs::msg::GnssFusionStatus> gnssFusionStatusMsgPtr;
-typedef std::unique_ptr<geometry_msgs::msg::PoseWithCovarianceStamped>
-  poseCovMsgPtr;
-typedef std::unique_ptr<geometry_msgs::msg::TransformStamped> transfMsgPtr;
-typedef std::unique_ptr<nav_msgs::msg::Odometry> odomMsgPtr;
-typedef std::unique_ptr<nav_msgs::msg::Path> pathMsgPtr;
-
-typedef std::unique_ptr<geographic_msgs::msg::GeoPoseStamped> geoPoseMsgPtr;
-typedef std::unique_ptr<sensor_msgs::msg::NavSatFix> navsatMsgPtr;
-
-typedef std::unique_ptr<zed_msgs::msg::ObjectsStamped> objDetMsgPtr;
-typedef std::unique_ptr<zed_msgs::msg::DepthInfoStamped> depthInfoMsgPtr;
-typedef std::unique_ptr<zed_msgs::msg::PlaneStamped> planeMsgPtr;
-typedef std::unique_ptr<visualization_msgs::msg::Marker> markerMsgPtr;
-
 typedef rclcpp::Service<std_srvs::srv::Trigger>::SharedPtr resetOdomSrvPtr;
 typedef rclcpp::Service<std_srvs::srv::Trigger>::SharedPtr resetPosTrkSrvPtr;
 typedef rclcpp::Service<zed_msgs::srv::SetPose>::SharedPtr setPoseSrvPtr;
@@ -184,7 +166,7 @@ typedef rclcpp::Service<zed_msgs::srv::StartSvoRec>::SharedPtr
 typedef rclcpp::Service<zed_msgs::srv::SetROI>::SharedPtr setRoiSrvPtr;
 typedef rclcpp::Service<std_srvs::srv::Trigger>::SharedPtr stopSvoRecSrvPtr;
 typedef rclcpp::Service<std_srvs::srv::Trigger>::SharedPtr pauseSvoSrvPtr;
-typedef rclcpp::Service<cob_srvs::srv::SetInt>::SharedPtr setSvoFramePtr;
+typedef rclcpp::Service<zed_msgs::srv::SetSvoFrame>::SharedPtr setSvoFramePtr;
 typedef rclcpp::Service<std_srvs::srv::Trigger>::SharedPtr resetRoiSrvPtr;
 typedef rclcpp::Service<robot_localization::srv::ToLL>::SharedPtr toLLSrvPtr;
 typedef rclcpp::Service<robot_localization::srv::FromLL>::SharedPtr fromLLSrvPtr;